fix routing and little fixes
